refactor!: Renamings in new handlers API
