+++ flow-server/src/main/java/com/vaadin/flow/server/streams/TransferContext.java
@@ -15,7 +15,6 @@
  */
 package com.vaadin.flow.server.streams;
 
-import java.io.Serializable;
 import java.util.Optional;
 
 import com.vaadin.flow.component.Component;
@@ -40,14 +39,14 @@ import com.vaadin.flow.server.VaadinSession;
  *            name of the file being transferred
  * @param owningElement
  *            the element that initiated the transfer
- * @param totalBytes
+ * @param contentLength
  *            the total number of bytes to be transferred or <code>-1</code> if
  *            total number is unknown in advance, e.g. when reading from an
  *            input stream
  */
 public record TransferContext(VaadinRequest request, VaadinResponse response,
         VaadinSession session, String fileName, Element owningElement,
-        long totalBytes) {
+        long contentLength) {
 
     /**
      * Get owner {@link Component} for this event.
