+++ flow-server/src/main/java/com/vaadin/flow/server/streams/UploadHandler.java
@@ -37,9 +37,7 @@ import org.slf4j.LoggerFactory;
 
 import com.vaadin.flow.dom.Element;
 import com.vaadin.flow.function.SerializableBiConsumer;
-import com.vaadin.flow.server.ElementRequestHandler;
 import com.vaadin.flow.server.HttpStatusCode;
-import com.vaadin.flow.server.TransferProgressListener;
 import com.vaadin.flow.server.VaadinRequest;
 import com.vaadin.flow.server.VaadinResponse;
 import com.vaadin.flow.server.VaadinSession;
