+++ flow-server/src/test/java/com/vaadin/flow/server/communication/StreamRequestHandlerTest.java
@@ -18,7 +18,7 @@ import com.vaadin.flow.component.UI;
 import com.vaadin.flow.dom.Element;
 import com.vaadin.flow.internal.CurrentInstance;
 import com.vaadin.flow.internal.StateNode;
-import com.vaadin.flow.server.ElementRequestHandler;
+import com.vaadin.flow.server.streams.ElementRequestHandler;
 import com.vaadin.flow.server.MockVaadinServletService;
 import com.vaadin.flow.server.MockVaadinSession;
 import com.vaadin.flow.server.ServiceException;
