+++ flow-server/src/test/java/com/vaadin/flow/server/streams/ClassDownloadHandlerTest.java
@@ -33,9 +33,6 @@ import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.dom.Element;
 import com.vaadin.flow.server.Command;
-import com.vaadin.flow.server.DownloadHandler;
-import com.vaadin.flow.server.DownloadEvent;
-import com.vaadin.flow.server.TransferProgressListener;
 import com.vaadin.flow.server.VaadinRequest;
 import com.vaadin.flow.server.VaadinResponse;
 import com.vaadin.flow.server.VaadinSession;
@@ -86,7 +83,7 @@ public class ClassDownloadHandlerTest {
                 new TransferProgressListener() {
                     @Override
                     public void onStart(TransferContext context) {
-                        Assert.assertEquals(-1, context.totalBytes());
+                        Assert.assertEquals(-1, context.contentLength());
                         Assert.assertEquals("download", context.fileName());
                         invocations.add("onStart");
                     }
@@ -103,7 +100,7 @@ public class ClassDownloadHandlerTest {
                     @Override
                     public void onComplete(TransferContext context,
                             long transferredBytes) {
-                        Assert.assertEquals(-1, context.totalBytes());
+                        Assert.assertEquals(-1, context.contentLength());
                         Assert.assertEquals(165000, transferredBytes);
                         Assert.assertEquals("download", context.fileName());
                         invocations.add("onComplete");
