+++ flow-server/src/test/java/com/vaadin/flow/server/streams/InputStreamDownloadHandlerTest.java
@@ -34,9 +34,6 @@ import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.dom.Element;
 import com.vaadin.flow.server.Command;
-import com.vaadin.flow.server.DownloadHandler;
-import com.vaadin.flow.server.DownloadEvent;
-import com.vaadin.flow.server.TransferProgressListener;
 import com.vaadin.flow.server.VaadinRequest;
 import com.vaadin.flow.server.VaadinResponse;
 import com.vaadin.flow.server.VaadinSession;
@@ -88,7 +85,7 @@ public class InputStreamDownloadHandlerTest {
         }, "download", new TransferProgressListener() {
             @Override
             public void onStart(TransferContext context) {
-                Assert.assertEquals(-1, context.totalBytes());
+                Assert.assertEquals(-1, context.contentLength());
                 Assert.assertEquals("download", context.fileName());
                 invocations.add("onStart");
             }
@@ -105,7 +102,7 @@ public class InputStreamDownloadHandlerTest {
             @Override
             public void onComplete(TransferContext context,
                     long transferredBytes) {
-                Assert.assertEquals(-1, context.totalBytes());
+                Assert.assertEquals(-1, context.contentLength());
                 Assert.assertEquals(165000, transferredBytes);
                 Assert.assertEquals("download", context.fileName());
                 invocations.add("onComplete");
