+++ flow-tests/test-root-context/src/main/java/com/vaadin/flow/uitest/ui/TransferProgressListenerView.java
@@ -16,24 +16,15 @@
 
 package com.vaadin.flow.uitest.ui;
 
-import java.io.ByteArrayInputStream;
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
 
-import com.vaadin.flow.component.html.Anchor;
-import com.vaadin.flow.component.html.AnchorTarget;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.Image;
-import com.vaadin.flow.component.html.Paragraph;
 import com.vaadin.flow.component.page.Push;
 import com.vaadin.flow.router.Route;
-import com.vaadin.flow.server.DownloadHandler;
-import com.vaadin.flow.server.HttpStatusCode;
-import com.vaadin.flow.server.VaadinSession;
+import com.vaadin.flow.server.streams.DownloadHandler;
 import com.vaadin.flow.server.streams.DownloadResponse;
-import com.vaadin.flow.uitest.servlet.ViewTestLayout;
 
 @Push
 @Route(value = "com.vaadin.flow.uitest.ui.TransferProgressListenerView")
